+++ electric-core/src/main/java/com/anton/electric/model/impl/Input380.java
@@ -34,7 +34,7 @@ public class Input380 extends AbstractComponent {
     private Output outputN;
 
     public Input380(String id, String name, double price) {
-        super(id, name, BigDecimal.valueOf(price));
+        super(id, name, price);
 
         this.inputL1 = new Input(this, ConnectorType.L);
         this.inputL2 = new Input(this, ConnectorType.L);
